feat(notifications): 423 - Adapt lazy loading of notifications for alerts / investigations.
