+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/base/mapper/QualityNotificationMapper.java
@@ -57,9 +57,7 @@ public class QualityNotificationMapper {
                 .edcNotificationId(notificationMessage.getEdcNotificationId())
                 .contractAgreementId(notificationMessage.getContractAgreementId())
                 .notificationReferenceId(notificationMessage.getNotificationReferenceId())
-                .isInitial(notificationMessage.getIsInitial())
                 .messageId(notificationMessage.getMessageId())
-                .edcUrl(notificationMessage.getEdcUrl())
                 .updated(notificationMessage.getUpdated())
                 .sendToName(notificationMessage.getSendToName())
                 .status(fromStatus(notificationMessage.getNotificationStatus()))
